+++ sources/@repo/docs/content/docs/bud.proxy/02-url-replacement.md
@@ -1,0 +1,8 @@
+---
+title: URL replacement
+description: Configure proxy server
+---
+
+import Replace from '@site/../../@roots/bud-api/docs/proxy/02-url-replacement.md'
+
+<Replace />
