+++ sources/@repo/docs/content/docs/bud.proxy/03-options.md
@@ -1,0 +1,13 @@
+---
+title: Options
+description: Configure proxy server
+---
+
+import Options from '@site/../../@roots/bud-api/docs/proxy/03-options.md'
+import Available from '@site/../../@roots/bud-api/docs/proxy/04-available-options.md'
+
+<Options />
+
+## Available options
+
+<Available />
