+++ sources/@repo/docs/content/docs/bud.proxy/index.md
@@ -1,0 +1,13 @@
+---
+title: bud.proxy
+description: Configure proxy server
+slug: index
+---
+
+import Usage from '@site/../../@roots/bud-api/docs/proxy/01-usage.md'
+
+Configure proxy server
+
+## Usage
+
+<Usage />
