+++ sources/@roots/bud-api/package.json
@@ -66,9 +66,11 @@
     "@roots/bud-extensions": "workspace:sources/@roots/bud-extensions",
     "@roots/bud-terser": "workspace:sources/@roots/bud-terser",
     "@skypack/package-check": "0.2.2",
+    "@types/http-proxy": "^1",
     "@types/node": "16.18.4",
     "copy-webpack-plugin": "11.0.0",
     "html-webpack-plugin": "5.5.0",
+    "http-proxy": "1.18.1",
     "http-proxy-middleware": "3.0.0-beta.0"
   },
   "dependencies": {
